Update AI add services to choose target pickup
Implementating restoration state of bot. In the beginning of the game it
should first find all weapons, then it should start collecting health,
armor, ammo and powerup pickups depending on it's needs and distance to
those pickups. It is needed to fix algorithm of choosing closest pickup
and algorithm of defining if pickup is not active to change target
pickup to another one.