Add FPSGameMode, add Bot pickup find logics
In FPSGameMode ID's of pickups are defined. Then with with help of
services FindPickupToTake and CalculatePickupNeeds Bot defines it's
target pickup and moves in the direction of it. Order of access block
were changed in all .h files. First goes interfaces (public then
protected) after interface goes private block. Added super 100 point
armor and implementation for it.