Add ControllerComponentsContainer
ControllerComponentsContainer is added to keep all components in one. It
is needed to ease adding of new component to both AI and player
controllers. Now it is enough to add new component to container. Also
added RespawnComponent to handle respawn logics after bot or player
dies. RespawnPoint replaces default PlayerStart because it's capsule
component collision is turned off and for implementing telefrag logics
it is needed to turn it on.

!!! It is needed to implement:
Score to determine which player should win
PlayersName should be changed because of Map ScoreBoard, because of it
in score board score changes for all bots.